Bugfix for bf_parse_json()
  In structures.h, var_type was redefined as uint8_t to save space in Var struct if
possible. This caused problems in the JSON parser, which defined two invalid types
(LIST_SENTINEL and MAP_SENTINEL) as -1 and -2 and inserted them into the bottom of
the stack to let the parser know when to stop parsing. Switching to unsigned caused
these to underflow and keep the loop running into invalid memory.

  Fix was to change the sentinel types to a valid type which won't be found in
JSON (TYPE_CLEAR). This seems to work well, but some more testing is warranted
before closing the issue for good.